develop-v2
:rocket: Adição de novas funcionalidades [Morphia Adicionado]

 - Adição de dependência jackson-module-kotlin para conseguir trabalhar melhor com valores nulos em kt
 - Trocando método findByUsernameAndPassword no qual retornava um User para existsByUsernameAndPassword no qual retorna um boolean
 - Trocando interface DomainEntity e todas as nomenclaturas dos ids de "_id" para "id"
 - Trocando tipagem dos params de Page de Long pra Int
 - Adição de nova exception UnprocessableEntity na Exception Handler
 - Remoção de alguns dos métodos no MongoExtensions e adição de outros
 - Trocando arquitetura GenericRepositoryMongoImpl usava KMongo agora é o Morphia
 - Trocando arquitetura CrudRepositoryService adição de algumas constraint validators
 - Adição de Indexes e Entity do Morphia para entidades
 - Adição de UNIQUE para email do User

+++ src/main/kotlin/org/edudev/arch/domain/Page.kt
@@ -2,6 +2,6 @@ package org.edudev.arch.domain
 
 
 class Page(
-    val first: Long = 0,
-    val last: Long = 10
+    val first: Int = 0,
+    val last: Int = 10
 )
